Fix bug where @2x.png would be used instead of .svg
When Sass compiles a media query, it places it at the end of the
current scope. Which means our @2x PNGs were being hoisted up,
after the SVG rules that were meant to be following them.

Wrapping the SVG rules in a `@media screen` block forces Sass
to *also* move the SVG rules to the end of the current scope,
*after* the PNG rules. Meaning browsers that support SVG image
backgrounds will now display those, rather than the @2x PNGs.